+++ src/Toml/FromValue.hs
@@ -8,40 +8,32 @@ Maintainer  : emertens@gmail.com
 Use 'FromValue' to define a transformation from some 'Value' to an application
 domain type.
 
-Use 'FromTable' to define transformations specifically from 'Table'. These
-instances are interesting because all top-level TOML values are tables,
-so these are the types that work for top-level deserialization.
-
-Use 'ParseTable' to help build 'FromTable' instances. It will make it easy to
-track which table keys have been used and which are left over.
+Use 'ParseTable' to help build 'FromValue' instances that match tables. It
+will make it easy to track which table keys have been used and which are left
+over.
 
 Warnings can be emitted using 'warning' and 'warnTable' (depending on what)
 context you're in. These warnings can provide useful feedback about
 problematic decodings or keys that might be unused now but were perhaps
 meaningful in an old version of a configuration file.
 
-"Toml.FromValue.Generic" can be used to derive instances of 'FromTable'
+"Toml.FromValue.Generic" can be used to derive instances of 'FromValue'
 automatically for record types.
 
 -}
 module Toml.FromValue (
     -- * Deserialization classes
     FromValue(..),
-    FromTable(..),
-    defaultTableFromValue,
 
     -- * Matcher
     Matcher,
     Result(..),
-    runMatcher,
-    withScope,
-    inKey,
-    inIndex,
     warning,
 
     -- * Table matching
     ParseTable,
     runParseTable,
+    parseTableFromValue,
     optKey,
     reqKey,
     warnTable,
@@ -66,9 +58,10 @@ import Data.String (IsString (fromString))
 import Data.Time (ZonedTime, LocalTime, Day, TimeOfDay)
 import Data.Word (Word8, Word16, Word32, Word64)
 import Numeric.Natural (Natural)
-import Toml.FromValue.Matcher (Matcher, Result(..), runMatcher, withScope, warning)
+import Toml.FromValue.Matcher (Matcher, Result(..), runMatcher, withScope, warning, inIndex, inKey)
 import Toml.Pretty (prettySimpleKey, prettyValue)
 import Toml.Value (Value(..), Table)
+import Toml.FromValue.ParseTable
 
 -- | Class for types that can be decoded from a TOML value.
 class FromValue a where
@@ -80,27 +73,20 @@ class FromValue a where
     listFromValue (Array xs) = zipWithM (\i v -> inIndex i (fromValue v)) [0..] xs
     listFromValue v = typeError "array" v
 
--- | Class for types that can be decoded from a TOML table.
-class FromValue a => FromTable a where
-    -- | Convert a 'Table' or report an error message
-    fromTable :: Table -> Matcher a
-
-instance (Ord k, IsString k, FromValue v) => FromTable (Map k v) where
-    fromTable t = Map.fromList <$> traverse f (Map.assocs t)
+instance (Ord k, IsString k, FromValue v) => FromValue (Map k v) where
+    fromValue (Table t) = Map.fromList <$> traverse f (Map.assocs t)
         where
             f (k,v) = (,) (fromString k) <$> inKey k (fromValue v)
-
-instance (Ord k, IsString k, FromValue v) => FromValue (Map k v) where
-    fromValue = defaultTableFromValue
-
--- | Derive 'fromValue' implementation from 'fromTable'
-defaultTableFromValue :: FromTable a => Value -> Matcher a
-defaultTableFromValue (Table t) = fromTable t
-defaultTableFromValue v = typeError "table" v
+    fromValue v = typeError "table" v
 
 -- | Report a type error
 typeError :: String {- ^ expected type -} -> Value {- ^ actual value -} -> Matcher a
 typeError wanted got = fail ("type error. wanted: " ++ wanted ++ " got: " ++ valueType got)
+
+-- | Used to derive a 'fromValue' implementation from a 'ParseTable' matcher.
+parseTableFromValue :: ParseTable a -> Value -> Matcher a
+parseTableFromValue p (Table t) = runParseTable p t
+parseTableFromValue _ v = typeError "table" v
 
 valueType :: Value -> String
 valueType = \case
@@ -202,47 +188,6 @@ instance FromValue LocalTime where
 instance FromValue Value where
     fromValue = pure
 
--- | A 'Matcher' that tracks a current set of unmatched key-value
--- pairs from a table.
---
--- Use 'optKey' and 'reqKey' to extract keys.
---
--- Use 'getTable' and 'setTable' to override the table and implement
--- other primitives.
-newtype ParseTable a = ParseTable (StateT Table Matcher a)
-    deriving (Functor, Applicative, Monad, Alternative, MonadPlus)
-
-instance MonadFail ParseTable where
-    fail = ParseTable . fail
-
--- | Lift a matcher into the current table parsing context.
-liftMatcher :: Matcher a -> ParseTable a
-liftMatcher = ParseTable . lift
-
--- | Run a 'ParseTable' computation with a given starting 'Table'.
--- Unused tables will generate a warning. To change this behavior
--- 'getTable' and 'setTable' can be used to discard or generate
--- error messages.
-runParseTable :: ParseTable a -> Table -> Matcher a
-runParseTable (ParseTable p) t =
- do (x, t') <- runStateT p t
-    case Map.keys t' of
-        []  -> pure x
-        [k] -> x <$ warning ("unexpected key: " ++ show (prettySimpleKey k))
-        ks  -> x <$ warning ("unexpected keys: " ++ intercalate ", " (map (show . prettySimpleKey) ks))
-
--- | Return the remaining portion of the table being matched.
-getTable :: ParseTable Table
-getTable = ParseTable get
-
--- | Replace the remaining portion of the table being matched.
-setTable :: Table -> ParseTable ()
-setTable = ParseTable . put
-
--- | Emit a warning at the current location.
-warnTable :: String -> ParseTable ()
-warnTable = ParseTable . lift . warning
-
 -- | Match a table entry by key if it exists or return 'Nothing' if not.
 optKey :: FromValue a => String -> ParseTable (Maybe a)
 optKey = optional . reqKey
@@ -250,55 +195,3 @@ optKey = optional . reqKey
 -- | Match a table entry by key or report an error if missing.
 reqKey :: FromValue a => String -> ParseTable a
 reqKey key = pickKey [Key key fromValue]
-
--- | Key and value matching function
---
--- @since 1.1.2.0
-data Alt a
-    = Key String (Value -> Matcher a) -- ^ pick alternative based on key match
-    | Else (Matcher a) -- ^ default case when no previous cases matched
-
--- | Take the first option from a list of table keys and matcher functions.
--- This operation will commit to the first table key that matches. If the
--- associated matcher fails, only that error will be propagated and the
--- other alternatives will not be matched.
---
--- If no keys match, an error message is generated explaining which keys
--- would have been accepted.
---
--- This is provided as an alternative to chaining multiple 'reqKey' cases
--- together with @('<|>')@ because that will generate one error message for
--- each unmatched alternative as well as the error associate with the
--- matched alternative.
---
--- @since 1.1.2.0
-pickKey :: [Alt a] -> ParseTable a
-pickKey xs =
- do t <- getTable
-    foldr (f t) (fail errMsg) xs
-    where
-        f t (Else m) _ = liftMatcher m
-        f t (Key k c) continue =
-            case Map.lookup k t of
-                Nothing -> continue
-                Just v ->
-                 do setTable $! Map.delete k t
-                    liftMatcher (inKey k (c v))
-
-        errMsg =
-            case xs of
-                []        -> "no alternatives"
-                [Key k _] -> "missing key: " ++ show (prettySimpleKey k)
-                _         -> "possible keys: " ++ intercalate ", " [show (prettySimpleKey k) | Key k _ <- xs]
-
--- | Update the scope with the message corresponding to a table key
---
--- @since 1.1.2.0
-inKey :: String -> Matcher a -> Matcher a
-inKey key = withScope ('.' : show (prettySimpleKey key))
-
--- | Update the scope with the message corresponding to an array index
---
--- @since 1.1.2.0
-inIndex :: Int -> Matcher a -> Matcher a
-inIndex i = withScope ("[" ++ show i ++ "]")